In HPIB you need to explicitly request a response read
